style: move context/credit info 2 prompt component
https://pi.dev/session/#338f800504032e429f0dac8c5b6cff4b
